fix: complete ak6 transport boundary cleanup
