+++ boundaries/atm-daemon/peer-http-adapter.toml
@@ -4,11 +4,11 @@ owner_crate_path = "atm_daemon"
 name = "PeerHttpAdapter"
 
 [public]
-trait = "HttpsMessageTransport"
+facade = "PeerHttpListenerSet"
 
 [implementation]
-type = "HttpsTransport"
-module = "atm_daemon::https_transport"
+type = "PeerHttpListenerSet"
+module = "atm_daemon::peer_http_listener"
 visibility = "pub(crate)"
 constructor = "pub(crate)"
 
@@ -16,7 +16,7 @@ constructor = "pub(crate)"
 roots = ["atm_daemon::composition::compose_runtime"]
 
 [ownership]
-io_owns = ["peer_http_https_socket_tls_adaptation", "peer_http_translation", "peer_ingress_authentication"]
+io_owns = ["peer_http_plain_socket_adaptation", "peer_http_translation", "peer_ingress_authentication"]
 io_forbidden = ["sqlite", "message_persistence", "delivery_queue", "retry_state", "recipient_routing", "nudge_emission"]
 
 [dependencies]
@@ -26,7 +26,7 @@ forbidden_edges = ["atm -> atm-daemon", "atm-graft -> atm-daemon"]
 
 [references]
 scope = "outside_owner_crate"
-forbidden = ["HttpsTransport", "HttpsListenerSet", "HttpsMessageTransport"]
+forbidden = ["PeerClientTransport", "peer_transport"]
 
 [contracts]
 request_types = ["HTTP WriteRequest"]
@@ -42,5 +42,5 @@ lint_rules = ["LINT-BOUNDARY-PEER-HTTP-ADAPTER-EDGES", "LINT-BOUNDARY-PEER-HTTP-
 review_gates = ["no_public_impl", "no_public_constructor", "no_storage_or_post_write_capability", "no_retired_frame_transport"]
 
 [status]
-state = "concrete_landed"
-notes = ["The live HTTP(S)/TLS adapter is distinct from the retired PeerClientTransport and peer_transport module."]
+state = "planned"
+notes = ["AK.4 owns the production trusted-LAN plain-HTTP receiver; the retired TLS adapter has no active boundary record."]
