+++ crates/atm-daemon/Cargo.toml
@@ -15,7 +15,6 @@ atm-core = { package = "agent-team-mail-core", path = "../atm-core", version = "
 atm-daemon-bootstrap = { path = "../atm-daemon-bootstrap", version = "1.4.0-beta-ai.38" }
 atm-runtime = { path = "../atm-runtime", version = "1.4.0-beta-ai.38" }
 atm-storage = { path = "../atm-storage", version = "1.4.0-beta-ai.38" }
-atm-peer-tls-interop = { path = "../atm-peer-tls-interop", version = "1.4.0-beta-ai.38" }
 arc-swap.workspace = true
 chrono = { version = "0.4", features = ["serde"] }
 fs2 = "0.4"
@@ -24,7 +23,6 @@ sc-observability.workspace = true
 sc-observability-types.workspace = true
 serde_json.workspace = true
 rustls.workspace = true
-sha2.workspace = true
 tracing.workspace = true
 
 [target.'cfg(unix)'.dependencies]
