+++ crates/atm-peer-tls-interop/Cargo.toml
@@ -13,7 +13,6 @@ description = "Provisioning and bounded curl mTLS interoperability fixture for A
 atm-core = { package = "agent-team-mail-core", path = "../atm-core", version = "1.4.0-beta-ai.38" }
 atm-storage = { path = "../atm-storage", version = "1.4.0-beta-ai.38" }
 rustls.workspace = true
-sha2.workspace = true
 
 [dev-dependencies]
 rcgen = "0.13"
