+++ crates/atm-storage/Cargo.toml
@@ -13,6 +13,8 @@ description = "Shared audited storage contract and canonical domain types for AT
 atm-error = { path = "../atm-error", version = "1.4.0-beta-ai.38" }
 serde.workspace = true
 serde_json.workspace = true
+rustls.workspace = true
+sha2.workspace = true
 chrono = { version = "0.4", features = ["serde"] }
 toml = "0.8"
 ulid.workspace = true
