+++ docs/adr/ADR-040-peer-authority-resolution.md
@@ -3,7 +3,7 @@
 | Field | Value |
 | --- | --- |
 | ID | ADR-040 |
-| Status | Proposed |
+| Status | Superseded by ADR-047 |
 | Scope | Repository-wide |
 | Relates to | ADR-034, ADR-035, Phase AI.25 |
