+++ docs/adr/ADR-041-end-to-end-peer-write-outcome.md
@@ -3,7 +3,7 @@
 | Field | Value |
 | --- | --- |
 | ID | ADR-041 |
-| Status | Accepted |
+| Status | Superseded by ADR-047 |
 | Scope | Repository-wide |
 | Relates to | ADR-032, ADR-034, ADR-035, Phase AI.26–AI.31 |
