+++ docs/adr/INDEX.md
@@ -42,13 +42,14 @@ crate-local ADR records that remain embedded in crate architecture documents.
 - [ADR-031 — Remote-Target Contract And Cross-Host Dispatch Boundary (superseded by ADR-034/035)](./ADR-031-remote-target-contract-and-cross-host-dispatch.md)
 - [ADR-032 — Unified Error Contract](./ADR-032-unified-error-contract.md)
 - [ADR-033 — HTTP Endpoint Contract](./ADR-033-http-endpoint-contract.md)
-- [ADR-034 — Minimal Cross-Host HTTPS Transport](./ADR-034-minimal-cross-host-https-transport.md)
-- [ADR-035 — Canonical Write Ingress And Host Routing](./ADR-035-canonical-write-ingress-and-host-routing.md)
+- [ADR-034 — Minimal Cross-Host HTTPS Transport (superseded by ADR-047)](./ADR-034-minimal-cross-host-https-transport.md)
+- [ADR-035 — Canonical Write Ingress And Host Routing (active ingress; transport wording superseded by ADR-047)](./ADR-035-canonical-write-ingress-and-host-routing.md)
 - [ADR-036 — Storage Boundary And Composition Topology](./ADR-036-storage-boundary-and-composition-topology.md)
 - [ADR-037 — Chat Address Identity](./ADR-037-chat-address-identity.md)
 - [ADR-039 — Python Graft Host Binding](./ADR-039-python-graft-host-binding.md)
-- [ADR-040 — Peer Authority Resolution](./ADR-040-peer-authority-resolution.md)
-- [ADR-041 — End-To-End Peer Write Outcome](./ADR-041-end-to-end-peer-write-outcome.md)
+- [ADR-040 — Peer Authority Resolution (superseded by ADR-047)](./ADR-040-peer-authority-resolution.md)
+- [ADR-041 — End-To-End Peer Write Outcome (superseded by ADR-047)](./ADR-041-end-to-end-peer-write-outcome.md)
+- ADR-047 — Direct trusted-LAN HTTP delivery (AK.4 decision; pending merge-forward)
 - [ADR-042 — SemVer Release And HTTP Compatibility](./ADR-042-semver-release-and-http-compatibility.md)
 - [ADR-043 — Hermes Graft Wake-up Ownership and Recovery](./ADR-043-hermes-graft-wake-up-ownership.md)
 - [ADR-044 — Public Verification Report Classification](./ADR-044-public-verification-report-classification.md)
