+++ docs/atm-storage/boundaries.md
@@ -2,6 +2,15 @@
 
 This document records shared storage-neutral contracts owned by `atm-storage`.
 
+## AK.6 TLS helper ownership
+
+atm_storage::tls owns the canonical certificate parsing, fingerprint
+normalization, and trusted-peer pinning helpers used while the legacy daemon
+adapter is retired. This is value validation and admission logic only: the
+storage crate owns no socket I/O, listener, sender, route, retry, or daemon
+lifecycle. The inactive atm-peer-tls-interop crate consumes these values for
+its bounded curl mTLS proof and has no active dependents.
+
 ## PeerConfigStore
 
 Canonical machine-readable boundary source:
